api/requirements.txt
azure-functions>=1.17.0
fastapi>=0.104.0
pydantic>=2.0.0
uvicorn>=0.24.0
